Исправлен обработчик кнопок

+++ app/src/main/java/com/example/calculator/MainActivity.java
@@ -5,29 +5,21 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.text.Editable;
+import android.widget.Button;
 import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
 
     private Calculator calculator;
     private TextView calculationText;
-    private final String NUMBER_ONE = "1";
-    private final String NUMBER_TWO = "2";
-    private final String NUMBER_THREE = "3";
-    private final String NUMBER_FOUR = "4";
-    private final String NUMBER_FIVE = "5";
-    private final String NUMBER_SIX = "6";
-    private final String NUMBER_SEVEN = "7";
-    private final String NUMBER_EIGHT = "8";
-    private final String NUMBER_NINE = "9";
-    private final String NUMBER_NULL = "0";
-    private final String SIGN_PLUS = "+";
-    private final String SIGN_MINUS = "-";
-    private final String SIGN_MULTIPLY = "*";
-    private final String SIGN_DIVISION = "/";
-    private final String SIGN_EQUALLY = "=";
     private TextView resultCalculationView;
     private final static String KEY_SAVE = "Calculation";
+    private final int[] numberButtonIds = new int[]{R.id.button_number_null, R.id.button_number_one,
+            R.id.button_number_three, R.id.button_number_four, R.id.button_number_five,
+            R.id.button_number_six, R.id.button_sign_delete, R.id.button_number_seven,
+            R.id.button_number_eight, R.id.button_number_nine, R.id.button_sign_equally,
+            R.id.button_sign_minus, R.id.button_sign_plus, R.id.button_sign_division,
+            R.id.button_sign_multiply, R.id.button_number_two};
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -36,7 +28,7 @@ public class MainActivity extends AppCompatActivity {
         calculationText = findViewById(R.id.calculation_text);
         resultCalculationView = findViewById(R.id.resultCalculation);
         calculator = new Calculator(calculationText, resultCalculationView);
-        buttonListener();
+        setNumberButtonListeners();
     }
 
     @Override
@@ -52,76 +44,21 @@ public class MainActivity extends AppCompatActivity {
         updateData();
     }
 
-    public void updateData(){
+    public void updateData() {
         calculationText.setText(calculator.getCalculationText().getText());
         resultCalculationView.setText(calculator.getResultCalculationView().getText());
     }
 
-    public void buttonListener() {
-        findViewById(R.id.button_number_null).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_NULL);
-        });
-
-        findViewById(R.id.button_number_one).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_ONE);
-        });
-
-        findViewById(R.id.button_number_two).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_TWO);
-        });
-
-        findViewById(R.id.button_number_three).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_THREE);
-        });
-
-        findViewById(R.id.button_number_four).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_FOUR);
-        });
-
-        findViewById(R.id.button_number_five).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_FIVE);
-        });
-
-        findViewById(R.id.button_number_six).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_SIX);
-        });
-
-        findViewById(R.id.button_number_seven).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_SEVEN);
-        });
-
-        findViewById(R.id.button_number_eight).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_EIGHT);
-        });
-
-        findViewById(R.id.button_number_nine).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(NUMBER_NINE);
-        });
-
-        findViewById(R.id.button_sign_division).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(SIGN_DIVISION);
-        });
-
-        findViewById(R.id.button_sign_multiply).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(SIGN_MULTIPLY);
-        });
-
-        findViewById(R.id.button_sign_minus).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(SIGN_MINUS);
-        });
-
-        findViewById(R.id.button_sign_plus).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(SIGN_PLUS);
-        });
-
-        findViewById(R.id.button_sign_equally).setOnClickListener(v -> {
-            calculator.addSymbolInCalculationText(SIGN_EQUALLY);
-        });
-
-        findViewById(R.id.button_sign_delete).setOnClickListener(v -> {
-            calculator.deleteLastSymbolInCalculationText();
-        });
-
+    private void setNumberButtonListeners() {
+        for (int i = 0; i < numberButtonIds.length; i++) {
+            findViewById(numberButtonIds[i]).setOnClickListener(v -> {
+                Button button = (Button) v;
+                if (button.equals(findViewById(R.id.button_sign_delete))) {
+                    calculator.deleteLastSymbolInCalculationText();
+                } else {
+                    calculator.addSymbolInCalculationText(button.getText().toString());
+                }
+            });
+        }
     }
-
 }
